backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1
hiredis==2.2.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Environment
python-dotenv==1.0.0

# Email
fastapi-mail==1.4.1

# File Processing
openpyxl==3.1.2
pandas==2.1.4

# Background Tasks
celery==5.3.4
flower==2.0.1

# API Documentation
fastapi-users==12.1.2

# Testing
factory-boy==3.3.0
faker==20.1.0

# Production
gevent==23.9.1